Adding modified versions of matlab niftiread
git-svn-id: https://svn.code.sf.net/p/apirl/code/trunk@365 4aa1ec02-d693-4363-b663-c1bee9069998

+++ matlab/imageProcessing/visualization/writeAnimatedGif.m
@@ -20,7 +20,7 @@ if ndims(volume) == 4
         %imind = uint8(round(double(volume(:,:,slices(n))./maxImage.*numColors)));
         %imind = double(volume(:,:,slices(n))./maxImage);
         % Write to the GIF File 
-        imwrite(volume(:,:,:,n),filename,'gif','WriteMode','append','DelayTime', frame_time); 
+        imwrite(double(volume(:,:,:,n)),filename,'gif','WriteMode','append','DelayTime', frame_time); 
     end
 else
     slices = find(mean(mean(volume))>0);
